requirements.txt
#streamlit==0.87.2
#pandas==1.3.2
plotly==5.2.2
matplotlib==3.4.3
Pillow==8.3.1




numpy>=1.9.2
matplotlib>=1.4.3
pandas==1.2.4
seaborn==0.10.1
streamlit==0.62.1
# wordcloud==1.6.0





